dates_exercise/requirements.txt
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
gspread==6.1.0
httplib2==0.22.0
idna==3.7
oauthlib==3.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
StrEnum==0.4.15
uritemplate==4.1.1
urllib3==2.2.1
